+++ libbpf-rs/tests/test_tc.rs
@@ -1,3 +1,4 @@
+use std::os::unix::io::AsFd as _;
 use std::os::unix::io::BorrowedFd;
 
 use serial_test::serial;
@@ -45,7 +46,7 @@ fn test_sudo_tc_basic_cycle() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -87,7 +88,7 @@ fn test_sudo_tc_attach_no_qdisc() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -112,7 +113,7 @@ fn test_sudo_tc_attach_basic() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -141,7 +142,7 @@ fn test_sudo_tc_attach_repeat() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -180,7 +181,7 @@ fn test_sudo_tc_attach_repeat() {
 fn test_sudo_tc_attach_custom() {
     bump_rlimit_mlock();
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -233,7 +234,7 @@ fn test_sudo_tc_attach_custom() {
 fn test_sudo_tc_detach_basic() {
     bump_rlimit_mlock();
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -280,7 +281,7 @@ fn test_sudo_tc_query() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder
@@ -352,7 +353,7 @@ fn test_sudo_tc_double_create() {
     bump_rlimit_mlock();
 
     let obj = get_test_object("tc-unit.bpf.o");
-    let fd = obj.prog("handle_tc").unwrap().fd();
+    let fd = obj.prog("handle_tc").unwrap().as_fd();
 
     let mut tc_builder = TcHookBuilder::new(fd);
     tc_builder